refactor: move buttons outside FiltersContainer
